More class hierarchies to extract similar functions for continuous envs
